+++ packages/x-charts-pro/src/ChartContainerPro/index.ts
@@ -1,4 +1,10 @@
 import '../typeOverloads';
 
+import type { ChartProApi as ChartProApiOriginal } from '../context/ChartProApi';
+
 export * from './ChartContainerPro';
-export type { ChartProApi } from './ChartProApi';
+
+/**
+ * @deprecated Use `ChartProApi` from `@mui/x-charts/context` instead.
+ */
+export type ChartProApi = ChartProApiOriginal;
