+++ packages/x-charts-pro/src/context/index.ts
@@ -1,1 +1,2 @@
-export * from './useChartApiContext';
+export * from './useChartProApiContext';
+export type { ChartProApi } from './ChartProApi';
