+++ packages/x-charts-pro/src/context/useChartProApiContext.ts
@@ -1,13 +1,13 @@
 import { useChartApiContext as useChartApiContextCommunity } from '@mui/x-charts/context';
-import { ChartProApi } from '../ChartContainerPro';
+import { ChartProApi } from './ChartProApi';
 
 /**
- * The `useChartApiContext` hook provides access to the chart API.
+ * The `useChartProApiContext` hook provides access to the chart API.
  * This is only available when the chart is rendered within a chart or a `ChartDataProvider` component.
  * If you want to access the chart API outside those components, you should use the `apiRef` prop instead.
  * @example
- * const apiRef = useChartApiContext<ChartProApi<'bar'>>();
+ * const apiRef = useChartProApiContext<ChartProApi<'bar'>>();
  */
-export function useChartApiContext<Api extends ChartProApi>() {
+export function useChartProApiContext<Api extends ChartProApi>() {
   return useChartApiContextCommunity<Api>();
 }
