+++ packages/x-charts/src/context/index.ts
@@ -5,3 +5,4 @@ export type {
   HighlightOptions,
 } from '../internals/plugins/featurePlugins/useChartHighlight';
 export * from './useChartApiContext';
+export type * from './ChartApi';
