+++ packages/x-charts/src/index.ts
@@ -23,6 +23,7 @@ export * from './SparkLineChart';
 export * from './Gauge';
 export * from './RadarChart';
 export * from './ChartsSurface';
-export * from './ChartContainer';
+export { ChartContainer } from './ChartContainer';
+export type { ChartContainerProps } from './ChartContainer';
 export * from './ChartDataProvider';
 export * from './Toolbar';
